refactor: self-review
