+++ apps/main/src/loan/components/PageCrvUsdStaking/DepositWithdraw/DeployButton.tsx
@@ -2,7 +2,7 @@ import { BigNumber } from 'bignumber.js'
 import { useCallback, useMemo } from 'react'
 import { styled } from 'styled-components'
 import { useConnection } from 'wagmi'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances.query'
 import { useStore } from '@/loan/store/useStore'
 import { Button } from '@ui/Button'
 import { t } from '@ui-kit/lib/i18n'
