+++ apps/main/src/loan/components/PageCrvUsdStaking/DepositWithdraw/DepositModule.tsx
@@ -2,7 +2,7 @@ import { BigNumber } from 'bignumber.js'
 import { useCallback } from 'react'
 import { useConnection } from 'wagmi'
 import { CRVUSD_ADDRESS, SCRVUSD_VAULT_ADDRESS } from '@/loan/constants'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances.query'
 import { networks } from '@/loan/networks'
 import { useStore } from '@/loan/store/useStore'
 import { type ChainId } from '@/loan/types/loan.types'
