+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/DistributionsBarChart.tsx
@@ -1,5 +1,5 @@
 import { BarChart, Bar, Cell, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts'
-import type { ScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue'
+import type { ScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue.query'
 import { Stack, Box } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
 import { formatDate } from '@ui/utils'
