+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/RevenueChartTooltip.tsx
@@ -1,5 +1,5 @@
 import type { YieldKeys } from '@/loan/components/PageCrvUsdStaking/types'
-import type { ScrvUsdYieldWithAverages } from '@/loan/entities/scrvusd-yield'
+import type { ScrvUsdYieldWithAverages } from '@/loan/entities/scrvusd-yield.query'
 import { formatDate } from '@ui/utils'
 import { t } from '@ui-kit/lib/i18n'
 import { ChartTooltipShell, ChartTooltipSeriesGroup, ChartTooltipSeriesRow } from '@ui-kit/shared/ui/Chart'
