+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/StatsStack.tsx
@@ -1,6 +1,6 @@
-import { useScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue'
-import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics'
-import { useScrvUsdYield } from '@/loan/entities/scrvusd-yield'
+import { useScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue.query'
+import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics.query'
+import { useScrvUsdYield } from '@/loan/entities/scrvusd-yield.query'
 import Grid from '@mui/material/Grid'
 import { t } from '@ui-kit/lib/i18n'
 import { Metric } from '@ui-kit/shared/ui/Metric'
