+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/index.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import { priceLineLabels } from '@/loan/components/PageCrvUsdStaking/Statistics/constants'
 import type { StatisticsChart, YieldKeys } from '@/loan/components/PageCrvUsdStaking/types'
-import { useScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue'
-import { useScrvUsdYield } from '@/loan/entities/scrvusd-yield'
+import { useScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue.query'
+import { useScrvUsdYield } from '@/loan/entities/scrvusd-yield.query'
 import { networks } from '@/loan/networks'
 import { useStore } from '@/loan/store/useStore'
 import { Stack, Card, CardHeader } from '@mui/material'
