+++ apps/main/src/loan/components/PageCrvUsdStaking/UserPosition/index.tsx
@@ -1,8 +1,8 @@
 import { BigNumber } from 'bignumber.js'
 import { useConnection } from 'wagmi'
 import { oneMonthProjectionYield, oneYearProjectionYield, isReady } from '@/loan/components/PageCrvUsdStaking/utils'
-import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
+import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics.query'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances.query'
 import { useStore } from '@/loan/store/useStore'
 import { Card, CardContent, CardHeader, Stack } from '@mui/material'
 import Grid from '@mui/material/Grid'
