+++ apps/main/src/loan/components/PageCrvUsdStaking/hooks/useScrvUsdDepositForm.ts
@@ -1,8 +1,8 @@
 import { useCallback, useMemo } from 'react'
 import { useConnection } from 'wagmi'
-import { useScrvUsdDepositIsApproved } from '@/loan/entities/scrvusd-deposit-approved.query'
+import { useScrvUsdDepositIsApproved } from '@/loan/entities/scrvusd-deposit-is-approved.query'
 import { useScrvUsdDepositMutation } from '@/loan/entities/scrvusd-deposit.mutation'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances.query'
 import { type ScrvUsdDepositForm, scrvUsdDepositFormValidationSuite } from '@/loan/entities/scrvusd.validation'
 import { type ChainId } from '@/loan/types/loan.types'
 import { useForm, useFormSync } from '@ui-kit/features/forms'
