+++ apps/main/src/loan/entities/scrvusd-deposit-is-approved.query.ts
