+++ apps/main/src/loan/entities/scrvusd-deposit.mutation.ts
@@ -7,7 +7,7 @@ import { t } from '@ui-kit/lib/i18n'
 import { rootKeys } from '@ui-kit/lib/model'
 import { type OnTransactionSuccess, useTransactionMutation } from '@ui-kit/lib/model/mutation/useTransactionMutation'
 import { formatNumber, waitForApproval } from '@ui-kit/utils'
-import { fetchScrvUsdDepositIsApproved } from './scrvusd-deposit-approved.query'
+import { fetchScrvUsdDepositIsApproved } from './scrvusd-deposit-is-approved.query'
 import { invalidateScrvUsdMutationQueries } from './scrvusd-mutation.helpers'
 import type { ScrvUsdDepositForm, ScrvUsdDepositMutation } from './scrvusd.validation'
 import { scrvUsdDepositMaxValidationSuite } from './scrvusd.validation'
