+++ apps/main/src/loan/entities/scrvusd-mutation.helpers.ts
@@ -1,12 +1,12 @@
 import { useConfig } from 'wagmi'
 import { CRVUSD_ADDRESS, SCRVUSD_VAULT_ADDRESS } from '@/loan/constants'
 import { invalidateScrvUsdDepositAllowance } from '@/loan/entities/scrvusd-deposit-allowance.query'
-import { invalidateScrvUsdDepositIsApproved } from '@/loan/entities/scrvusd-deposit-approved.query'
-import { invalidateScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
 import type { ChainId } from '@/loan/types/loan.types'
 import type { Address } from '@primitives/address.utils'
 import { invalidateTokenBalances } from '@ui-kit/hooks/useTokenBalance'
 import type { UserChainParams } from '@ui-kit/lib/model'
+import { invalidateScrvUsdDepositIsApproved } from './scrvusd-deposit-is-approved.query'
+import { invalidateScrvUsdUserBalances } from './scrvusd-userBalances.query'
 
 export const invalidateScrvUsdMutationQueries = async ({
   chainId,
