+++ apps/main/src/loan/entities/scrvusd-statistics.query.ts
