+++ apps/main/src/loan/entities/scrvusd-yield.query.ts
